// src/components/forms/Form/interface.ts
export type InternalNamePath = (string | number)[];
export type NamePath = string | number | InternalNamePath;

// eslint-disable-next-line @typescript-eslint/no-explicit-any
export type StoreValue = any;
export type Store = Record<string, StoreValue>;

export interface Meta {
  touched: boolean;
  validating: boolean;
  errors: string[];
  warnings: string[];
  name: InternalNamePath;
  rules: RuleObject[];
}

export interface InternalFieldData extends Meta {
  value: StoreValue;
}

// Used by `setFields` config
export interface FieldData extends Partial<Omit<InternalFieldData, 'name'>> {
  name: NamePath;
}

export type FieldsErrors = Record<string, Meta['errors']>;
// ...
export type RuleObject = AggregationRule | ArrayRule;

export type Rule = RuleObject | RuleRender;
export interface ValidateErrorEntity<Values = unknown> {
  values: Values;
  errorFields: { name: InternalNamePath; errors: string[] }[];
  outOfDate: boolean;
}
// ...
export interface FieldError {
  name: InternalNamePath;
  errors: string[];
  warnings: string[];
}
// ...
export type ValidateFields<Values = unknown> = (
  nameList?: NamePath[],
) => Promise<Values>;
// ...
export type SetErrorsOptions = {
  prefix?: string;
};

// Only return partial when type is not any
type RecursivePartial<T> = T extends object
  ? {
      [P in keyof T]?: T[P] extends (infer U)[]
        ? RecursivePartial<U>[]
        : T[P] extends object
        ? RecursivePartial<T[P]>
        : T[P];
    }
  : // eslint-disable-next-line @typescript-eslint/no-explicit-any
    any;

// eslint-disable-next-line @typescript-eslint/no-explicit-any
export interface FormInstance<Values = any> {
  // Origin Form API
  getFieldValue: (name: NamePath) => StoreValue;
  getFieldsValue: (() => Values) &
    ((
      nameList: NamePath[] | true | null,
      filterFunc?: (meta: Meta) => boolean,
      // eslint-disable-next-line @typescript-eslint/no-explicit-any
    ) => any);
  getFieldError: (name: NamePath) => string[];
  getFieldsError: (nameList?: NamePath[]) => FieldError[];
  getFieldWarning: (name: NamePath) => string[];
  isFieldsTouched: ((
    nameList?: NamePath[],
    allFieldsTouched?: boolean,
  ) => boolean) &
    ((allFieldsTouched?: boolean) => boolean);
  isFieldTouched: (name: NamePath) => boolean;
  isFieldValidating: (name: NamePath) => boolean;
  isFieldsValidating: (nameList: NamePath[]) => boolean;
  resetFields: (fields?: NamePath[]) => void;
  setError: (
    name: NamePath,
    error: Meta['errors'],
    options?: SetErrorsOptions,
  ) => void;
  setErrors: (error: FieldsErrors, options?: SetErrorsOptions) => void;
  setFields: (fields: FieldData[]) => void;
  setFieldValue: (name: NamePath, value: unknown) => void;
  setFieldsValue: (values: RecursivePartial<Values>) => void;
  validateFields: ValidateFields<Values>;

  // New API
  submit: () => void;
}
// ...
type ValidateMessage = string | (() => string);
export interface ValidateMessages {
  default?: ValidateMessage;
  required?: ValidateMessage;
  enum?: ValidateMessage;
  whitespace?: ValidateMessage;
  date?: {
    format?: ValidateMessage;
    parse?: ValidateMessage;
    invalid?: ValidateMessage;
    max?: ValidateMessage;
    min?: ValidateMessage;
    range?: ValidateMessage;
  };
  types?: {
    string?: ValidateMessage;
    method?: ValidateMessage;
    array?: ValidateMessage;
    object?: ValidateMessage;
    number?: ValidateMessage;
    date?: ValidateMessage;
    boolean?: ValidateMessage;
    integer?: ValidateMessage;
    float?: ValidateMessage;
    regexp?: ValidateMessage;
    email?: ValidateMessage;
    url?: ValidateMessage;
    hex?: ValidateMessage;
  };
  string?: {
    len?: ValidateMessage;
    min?: ValidateMessage;
    max?: ValidateMessage;
    range?: ValidateMessage;
  };
  number?: {
    len?: ValidateMessage;
    min?: ValidateMessage;
    max?: ValidateMessage;
    range?: ValidateMessage;
  };
  array?: {
    len?: ValidateMessage;
    min?: ValidateMessage;
    max?: ValidateMessage;
    range?: ValidateMessage;
  };
  pattern?: {
    mismatch?: ValidateMessage;
  };
}